requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
selenium==4.12.0
whisper==1.0
openai==0.27.8
python-dotenv==1.0.0
pydub==0.25.1
flask==2.3.2